fix(visualizer): separate per-frame and per-pixel preset state
MilkDrop keeps the per-frame and per-vertex expression blocks in separate variable universes (m_pf_eel/m_pv_eel), so a per-pixel write cannot clobber the per-frame spring state. Orynivo shared one slot array, so Royal Mashup (13)'s per-pixel vortex points overwrote the per-frame spring (q4/q5/q8) and the render diverged from Winamp. The per-pixel block now runs on its own array, seeded from the per-frame state each frame, and slot values are double like ns-eel2's EEL_F.

Also align the comp shader's rad/ang with MilkDrop's UvToMathSpace on the CPU interpreter (the GPU emitter already used it; CompShaderPolarTests pins it), publish the fShader hue amount, restore the small diamond centre darkening, interpolate custom-wave line colours, give thick custom-wave dots their two-by-two footprint, and fix the aspecty regression test expectation. verify-all passes.

+++ Orynivo.Core.Tests/PresetCompilerTests.cs
@@ -21,7 +21,7 @@ public sealed class PresetCompilerTests
 
         Assert.True(program.IsEmpty);
         Assert.Empty(program.Variables);
-        program.Execute(new float[4]);
+        program.Execute(new double[4]);
     }
 
     /// <summary>An assignment writes the variable's slot.</summary>
@@ -30,7 +30,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = 0.25;");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         program.Execute(slots);
 
         Assert.Single(program.Variables);
@@ -43,7 +43,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = 2 + 3 * 4; q2 = (2 + 3) * 4; q3 = 10 - 2 - 3;");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         program.Execute(slots);
 
         Assert.Equal(14f, slots[program.IndexOf("q1")]);
@@ -57,7 +57,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = -3 + 1; q2 = !0; q3 = !2; q4 = 7 % 3; q5 = -7 % 3;");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         program.Execute(slots);
 
         Assert.Equal(-2f, slots[program.IndexOf("q1")]);
@@ -74,7 +74,7 @@ public sealed class PresetCompilerTests
         var program = PresetCompiler.Compile(
             "q1 = 2 > 1; q2 = 1 > 2; q3 = 1 == 1; q4 = 1 != 1; q5 = 1 < 2 && 2 < 3; q6 = 0 || 0; q7 = 2 >= 2;");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         program.Execute(slots);
 
         Assert.Equal(1f, slots[program.IndexOf("q1")]);
@@ -92,7 +92,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = bass > 0.5 ? 1 : 0; q2 = if(mid, 2, 3);");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         slots[program.IndexOf("bass")] = 0.75f;
         slots[program.IndexOf("mid")] = 0f;
         program.Execute(slots);
@@ -109,7 +109,7 @@ public sealed class PresetCompilerTests
             "q1 = sin(0); q2 = cos(0); q3 = sqrt(9); q4 = abs(-4); q5 = floor(1.7); q6 = ceil(1.2);"
             + " q7 = int(-1.7); q8 = min(2, 3); q9 = max(2, 3); q10 = pow(2, 3); q11 = sign(-2); q12 = log10(100);");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         program.Execute(slots);
 
         Assert.Equal(0f, slots[program.IndexOf("q1")], 5);
@@ -132,7 +132,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = bass * 2; q2 = q1 + time;");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         slots[program.IndexOf("bass")] = 0.25f;
         slots[program.IndexOf("time")] = 10f;
         program.Execute(slots);
@@ -147,7 +147,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = pi;");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         program.Execute(slots);
 
         Assert.Equal(MathF.PI, slots[program.IndexOf("q1")], 5);
@@ -159,7 +159,7 @@ public sealed class PresetCompilerTests
     public void Execute_KeepsRandomValuesInRange()
     {
         var program = PresetCompiler.Compile("q1 = rand(2);");
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
 
         for (var attempt = 0; attempt < 50; attempt++)
         {
@@ -174,7 +174,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = 1; // a comment\n;; q2 = 2;");
 
-        var slots = new float[program.Variables.Count];
+        var slots = new double[program.Variables.Count];
         program.Execute(slots);
 
         Assert.Equal(1f, slots[program.IndexOf("q1")]);
@@ -187,7 +187,7 @@ public sealed class PresetCompilerTests
     {
         var program = PresetCompiler.Compile("q1 = 1; q2 = 2;");
 
-        Assert.Throws<ArgumentException>(() => program.Execute(new float[1]));
+        Assert.Throws<ArgumentException>(() => program.Execute(new double[1]));
     }
 
     /// <summary>Author mistakes are reported with a position.</summary>
